Moved ContactDataOps, ContactPersonOps, Persons and ContactEntries classes to their own package.
Created packages for adapters and service.
Created PersonAdapter to populate an appropriate ListView for the data.
Testing the addition of a person data to the SQL database.

TO-DO:

Integrate Activity with helper classes.
Test functionality of record writing and reading.
Set-up ArrayAdapter replacement.
Begin coding the service that will underpin the application.