Service implementation has begun. It can curently be launched from the debug menu, and will run until a kill argument is sent, at which point it will terminate itself.
Work has been ongoing on the locationlistener logic for GPS positioning. Last location provisioning is implemented but needs testing. Location updates pending.

NotificationManager implemented. This will allow the service to issue notifications to the Android notification system to provide a variety of data to the user.

To do:
Utilise the lon/lat data from the API to provide a visual location on a map, for use in setting geofencing for dangerous locations.
Implement service control in the end-user GUI
Implement a call to Preferences as established in the PrefsActivity for use in the service itself (currently it uses hardcoded defaults).
Implement a SQL DB for PrefsActivity to store user settings.
Implement call, message and IM logic, and tie it in to the existing internal contacts DB.