SMS Message Tasking implemented. Needs testing.
To do:
Utilise the lon/lat data from the API to provide a visual location on a map, for use in setting geofencing for dangerous locations.
Implement service control in the end-user GUI
Implement a call to Preferences as established in the PrefsActivity for use in the service itself (currently it uses hardcoded defaults).
Implement a SQL DB for PrefsActivity to store user settings.
Implement call, message and IM logic, and tie it in to the existing internal contacts DB.